backend/requirements.txt
Django==5.0
django-filter==23.5
djangorestframework==3.14.0
django-cors-headers==4.3.1
Pillow==10.1.0
qrcode==7.4.2
python-dotenv==1.0.0
django-apscheduler==0.6.2
dj-database-url==2.1.0
whitenoise==6.6.0
gunicorn==21.2.0
psycopg2-binary==2.9.9